DEV-2277 Wait for all events for a run
Previously the snpcheck was only processing the tertiary events from any given
run. We only want to really act on the tertiary events but we also want to
propagate the others so that the archiver can handle the files described
by them too.

The implementation is coloured by a preference to avoid introducing additional
threads. It's also slightly more complex because we have to be aware there may
be event replays happening, so each event must be explicitly checked before we
can be sure we have a full set. The implementation will be prone to problems if
the executing pod is restarted after the some but before all three events
arrive; in that case we will need to notice and replay the events.

I am aware this isn't the nicest-looking code, comments for improvement
requested.