DEV-2277 Simplify event handling
I've reverted the change from the last commit and simplified the logic so that
essentially the snpcheck will be done three times for every run/type: one for
each event received, each of which has one of the contexts we recognise
(currently germline, secondary, tertiary). Shallow runs are still exempted
however.

This should be fine from a processing perspective as it will just introduce
extraneous snpcheck operations. It is worth it for the simplification in the
event handling, which ultimately can now be moved to the archiver if needed.