DEV-2277 Handle QC failure case
QC failures must still have their events passed through to the archiver so add
handling for that.

Also add test cases to cover all combinations of contexts and types since we
there are downstream consequences to the logic here and it is getting hard
to keep track of what should do what.

Noticed there was still a deployment for the development environment which is
now in the deployment-manager project and against the staging environment. We
will have to work out how to do an automatic deploy using that if we want to
have it.