DEV-2759 Re-queue buffered events on failure
Prior to this change any exception when a buffered event was reprocessed would
essentially be swallowed as the handler would log the error only and the caller
would have no way to know something had happened.

Instead we now re-throw any exception as an unchecked exception and re-buffer
the request in this situation.