DEV-2948 Validate research runs if there is snpchecked diagnostic
Check for the most recent diagnostic run and if it has not failed on the snpcheck, publish validated.

If no diagnostic runs, throw an error as that should be impossible. Warn if the most recent diagnostic
failed snpcheck, as the diagnostic run will be investigated and likely require intervention already.

Also remove a try catch around the whole block, its handled in the subscriber (makes unit testing the
thrown exception easier)