INFRA-1330 Adjust api update logic
Previously, if a run had a fail status with QC failure, then the API status would never be updated to validated even if a consecutive snpcheck passes.

However, this logic applied for both HealthCheck QC failure and SnpCheck QC failure. The latter does not make much sense. This is because this failure status means the snpcheck failed, but if we then do a new run and this succeeds, we would expect the api to be updated to validated.

This PR fixes this, and leaves the healthcheck logic in tact.